Add cover script in package.json and fix code coverage not reporting for the secondary endpoint.
