+++ package.json
@@ -9,7 +9,8 @@
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "test:lib": "ng test my-lib",
-    "test:ci": "ng test --no-watch --browsers=ChromeHeadless"
+    "test:ci": "ng test --no-watch --browsers=ChromeHeadless",
+    "cover": "ng test --no-watch --code-coverage"
   },
   "private": true,
   "dependencies": {
@@ -30,6 +31,7 @@
     "@angular/cli": "~14.0.1",
     "@angular/compiler-cli": "^14.0.0",
     "@types/jasmine": "~4.0.0",
+    "@types/webpack-env": "^1.18.0",
     "jasmine-core": "~4.1.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
