+++ projects/my-lib/secondary/src/test.ts
@@ -1,0 +1,5 @@
+import { initTestBed } from '../../test-shared';
+
+initTestBed();
+const context = require.context('./', true, /\.spec\.ts$/);
+context.keys().forEach(context);
