+++ projects/my-lib/secondary/tsconfig.spec.json
@@ -1,0 +1,7 @@
+/* To learn more about this file see: https://angular.io/config/tsconfig. */
+{
+  "extends": "../tsconfig.spec.json",
+  "files": [
+    "src/test.ts"
+  ]
+}
