+++ projects/my-lib/tsconfig.lib.json
@@ -6,7 +6,9 @@
     "declaration": true,
     "declarationMap": true,
     "inlineSources": true,
-    "types": []
+    "types": [
+      "webpack-env"
+    ]
   },
   "exclude": [
     "test.ts",
